# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Delivery",
        "description": "API del Sistema de Gestión de Delivery",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "Productos obtenidos correctamente."
                    },
                    "404": {
                        "description": "No hay productos disponibles para mostrar."
                    },
                    "500": {
                        "description": "Error al obtener los productos."
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "48b59f126c849beb4d4de4ff83b7945f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categoria_id",
                                    "nombre",
                                    "precio"
                                ],
                                "properties": {
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hamburguesa Doble",
                                        "maxLength": 100
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Doble carne, queso cheddar y panceta",
                                        "nullable": true
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4500.5
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://images.unsplash.com/photo-1568901346375-23c9450c58cd?w=500",
                                        "nullable": true
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente."
                    },
                    "422": {
                        "description": "Error de validación."
                    },
                    "500": {
                        "description": "Error interno al crear el producto."
                    }
                }
            }
        },
        "/api/productos-desactivados": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos desactivados",
                "operationId": "d7f6e10c37936332c2d31aae9fce3f57",
                "responses": {
                    "200": {
                        "description": "Productos desactivados obtenidos correctamente."
                    },
                    "404": {
                        "description": "No hay productos desactivados para mostrar."
                    },
                    "500": {
                        "description": "Error al obtener los productos desactivados."
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por ID",
                "operationId": "672fde814ef8716ee902835cc96ae595",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido correctamente."
                    },
                    "404": {
                        "description": "Producto no encontrado."
                    },
                    "500": {
                        "description": "Error al obtener el producto."
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto por ID",
                "operationId": "20b20d8e59992eb88abad7dcef3db6d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "activo"
                                ],
                                "properties": {
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hamburguesa Triple",
                                        "maxLength": 100
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Triple carne y queso",
                                        "nullable": true
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5500
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://images.unsplash.com/photo-1568901346375-23c9450c58cd?w=500",
                                        "nullable": true
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente."
                    },
                    "404": {
                        "description": "Producto no encontrado."
                    },
                    "422": {
                        "description": "Error de validación."
                    },
                    "500": {
                        "description": "Error interno del servidor."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto por ID",
                "operationId": "4c0c7680428675134afd4d793f934f67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente."
                    },
                    "404": {
                        "description": "Producto no encontrado."
                    },
                    "500": {
                        "description": "Error interno del servidor."
                    }
                }
            }
        },
        "/api/productos/{id}/restore": {
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Restaurar un producto eliminado",
                "operationId": "2a6854a610367c3f8d2310e72a81b3c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a restaurar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto restaurado correctamente."
                    },
                    "404": {
                        "description": "Producto no encontrado."
                    },
                    "500": {
                        "description": "Error interno del servidor."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}